Fix to look up case-insensitive reference link
Current markedj converts the key of link definition to lower case,
but it dosen't convert the key of reference link. It causes the key
not to look up. See following Markdown:

```markdown
[Foo]

[Foo]: http://example.com
```

It yields to `<p>[Foo]</p>`, but the result we expected is
`<p><a href="http://example.com">Foo</a></p>` because the key `Foo`
can't be looked up.

I fixed the key of reference link is converted to lower case when
looking it up from definitions.